- https://github.com/UbaidurRehman1/spring-boot-with-thymeleaf/issues/24
